requirements.txt
pika
mysql-connector-python
PyJWT